RecipeProject2/RecipeApp/requirements.txt
Django==4.1.13
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
uuid==1.30